reverse with view

Reverse a multidimensional array with a view instead of a copy.

NDTools.reverse_view — Function

reverse_view(A::AbstractArray{T, N}; 
             dims=ntuple(identity, Val(N))) where {T, N}

Creates a view of an array A which reverses all dimensions in dims. Per default, dims is set to reverse all dimensions! It is equivalent to reverse(A, dims=dims) but instead uses a view.

Examples

julia> A = [1 2 3; 4 5 6]
2×3 Matrix{Int64}:
 1  2  3
 4  5  6

julia> reverse_view(A, dims=1)
2×3 view(::Matrix{Int64}, 2:-1:1, 1:3) with eltype Int64:
 4  5  6
 1  2  3

julia> reverse(A, dims=1)
2×3 Matrix{Int64}:
 4  5  6
 1  2  3

julia> reverse_view(A, dims=(1,2))
2×3 view(::Matrix{Int64}, 2:-1:1, 3:-1:1) with eltype Int64:
 6  5  4
 3  2  1

julia> reverse(A, dims=(1,2))
2×3 Matrix{Int64}:
 6  5  4
 3  2  1