Dimensionality Functions

NDTools.expand_dims — Function

expand_dims(x, ::Val{N})

Expands the dimensions of an array to a given number of dimensions.

Examples The result is a 5D array with singleton dimensions at the end

julia> expand_dims(ones((1,2,3)), Val(5))
1×2×3×1×1 Array{Float64, 5}:
[:, :, 1, 1, 1] =
 1.0  1.0

[:, :, 2, 1, 1] =
 1.0  1.0

[:, :, 3, 1, 1] =
 1.0  1.0

expand_dims(x::AbstractArray{T, N}, dims::Vararg{Int, M})  where {T, N, M}

Insert singleton dimensions at the position of dims into x. Based on a reshape operation.

Examples

julia> expand_dims(zeros((2,2)), 1) |> size
(1, 2, 2)

julia> expand_dims(zeros((2,2)), 2) |> size
(2, 1, 2)

julia> expand_dims(zeros((2,2)), 3) |> size
(2, 2, 1)

julia> expand_dims(zeros((2,2)), 1,3,4) |> size
(1, 2, 1, 1, 2)

NDTools.expand_size — Function

expand_size(sz,sz2)

Expands a size tuple sz with the sizes as given in the tuple sz2 for positions which do not exist in sz. Typically one wants to

Example:

julia> sz = expand_size((1,2,3),(4,5,6,7,8,9))
Base.Generator{UnitRange{Int64}, var"#7#8"{Tuple{Int64, Int64, Int64}, NTuple{6, Int64}, Int64}}(var"#7#8"{Tuple{Int64, Int64, Int64}, NTuple{6, Int64}, Int64}((1, 2, 3), (4, 5, 6, 7, 8, 9), 3), 1:6)

julia> Tuple(sz)
(1, 2, 3, 7, 8, 9)

NDTools.select_sizes — Function

select_sizes(x::AbstractArray, dim)

Additional size method to access the size at several dimensions in one call. Keep singleton dimensions.

Examples

julia> x = ones((2,4,6,8, 10));

julia> select_sizes(x, (2,3))
(1, 4, 6, 1, 1)

julia> select_sizes(x, 5)
(1, 1, 1, 1, 10)

julia> select_sizes(x, (5,))
(1, 1, 1, 1, 10)

NDTools.select_sizes_squeeze — Function

select_sizes_squeeze(x::AbstractArray, dim)

Additional size method to access the size at several dimensions in one call. Remove singleton dimensions.

See also select_sizes which does not remove singleton dimensions.

Examples

julia> select_sizes_squeeze(randn((5,6,7)), (2,3))
(6, 7)

julia> select_sizes_squeeze(randn((5,6,7)), 2)
(6,)